Move Tango conversions to AttributeProxy
